Mass commit or remaining assignments and final assesment

+++ CSE17/project4/Sort.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
  * Sort
  */
 public class Sort {
-    public static long iterations = 0;
+    public static Integer iterations = 0;
 
     /**
      * @return the iterations of most recently used sorting algorithm
